NEW: User - Create & Edit

+++ src/main/java/Hieu_iceTea/FoodMate_Spring/controller/dashboard/UserController.java
@@ -6,6 +6,7 @@ import Hieu_iceTea.FoodMate_Spring.utilities.storage.service.StorageService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -59,8 +60,15 @@ public class UserController {
         return "dashboard/user/create-edit";
     }
 
-    @PostMapping(path = {"", "/"})
-    public String store(@ModelAttribute User user, @RequestParam("image_file") MultipartFile file) {
+    @PostMapping(path = {"/create/", "/create"})
+    public String store(@ModelAttribute User user,
+                        BindingResult bindingResult,
+                        @RequestParam("image_file") MultipartFile file) {
+
+        //Xử lý Validating-Form
+        if (bindingResult.hasErrors()) {
+            return "dashboard/user/create-edit";
+        }
 
         //Xử lý file
         if (!file.isEmpty()) {
@@ -85,13 +93,23 @@ public class UserController {
         return "dashboard/user/create-edit";
     }
 
-    @PostMapping(path = {"/{id}/", "/{id}"})
-    public String update(@ModelAttribute User user, @RequestParam("image_file") MultipartFile file, @RequestParam("image_old") String fileName_old) {
+    @PostMapping(path = {"/{id}/edit/", "/{id}/edit"})
+    public String update(@ModelAttribute User user,
+                         BindingResult bindingResult,
+                         @RequestParam("image_file") MultipartFile file,
+                         @RequestParam("image_old") String fileName_old) {
+
+        //Xử lý Form-Validation
+        if (bindingResult.hasErrors()) {
+            return "dashboard/user/create-edit";
+        }
 
         //Xử lý file
         if (!file.isEmpty()) {
             // 01. Xóa file cũ:
-            storageService.delete(fileName_old, _imagePath);
+            if (fileName_old != null && !fileName_old.isBlank()) {
+                storageService.delete(fileName_old, _imagePath);
+            }
 
             // 02. Lưu file mới:
             String fileName =  storageService.store(file, _imagePath);
@@ -110,7 +128,10 @@ public class UserController {
     public String delete(@PathVariable int id) {
 
         // 01. Xóa file:
-        storageService.delete(userService.findById(id).getImage(), _imagePath);
+        String fileName_old = userService.findById(id).getImage();
+        if (fileName_old != null && !fileName_old.isBlank()) {
+            storageService.delete(fileName_old, _imagePath);
+        }
 
         // 02. Xóa bản ghi database
         userService.deleteById(id);
